After The Deadline: ATD Core: Fix XSS issue described here: https://hackerone.com/reports/72003
This splits leading and trailing text in a processed text node into new sibling text nodes around the "underlined" text.

Previous behavior was to create a new HTML element with the whole thing thereby executing scripts in the surrounding text (in some browsers).

This limits the vector to the matched text which _should_ be safe due to tokenization.

The change is limited to the non-IE code path because I wasn't able to induce XSS via this method during IE testing.

Merges r120816-wpcom.